Large breaking refactor
This commit improves the codebase in multiple ways.
- Create new Config class to handle user configuration. For now only saves microphone and audio format preferences.
- Change default audio format to mp3 128k.
- Add ability to choose an audio format in setup.
- Now both microphone and audio format are written to the config file
- Add new console argument "fix". If the app fails in some way before writing the final audio file, this command can be used to traverse the SimpleRecorder folder and merge channels into final files.

+++ src/SimpleRecorder/Program.cs
@@ -2,17 +2,19 @@
 using NAudio.Wave;
 using NAudio.Wave.SampleProviders;
 using System.Runtime.InteropServices;
+using System.Text.RegularExpressions;
+using System.Windows.Markup;
 using static System.Console;
 
 namespace SimpleRecorder;
 class Program
 {
 	static string outputFolder = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.MyMusic), "SimpleRecorder");
-	static string configFolder = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData), "SimpleRecorder");
-	static string config = Path.Combine(configFolder, ".config");
+	
 	static string stereopath = string.Empty;
 	static string inpath = string.Empty;
-	static string finalpath = string.Empty;
+	static string iso = string.Empty;
+
 	static ISampleProvider silenceprovider;
 
 	static DateTime start = DateTime.Now;
@@ -24,6 +26,8 @@ class Program
 	static WaveInEvent waveIn;
 	static WaveOutEvent silenceOut;
 	static WasapiLoopbackCapture loopback;
+
+	static Config config;
 
 	static bool isRecording = false;
 
@@ -51,7 +55,7 @@ class Program
 			case CtrlType.CTRL_SHUTDOWN_EVENT:
 			case CtrlType.CTRL_CLOSE_EVENT:
 			default:
-				CloseAndWrite();
+				CloseAndWrite(config);
 				return false;
 		}
 	}
@@ -61,17 +65,62 @@ class Program
 		_handler += new EventHandler(Handler);
 		SetConsoleCtrlHandler(_handler, true);
 
-		(int mic_id, string name) id_info;
 		var cts = new CancellationTokenSource();
 
-		if (args.Length == 1 && args[0] == "setup")
+		switch (args.Length)
 		{
-			id_info = Setup();
+			case 1:
+				switch (args[0])
+				{
+					case "setup" or "Setup":
+						config = Config.Setup();
+						break;
+					case "fix" or "Fix":
+						config = Config.FromFile();
+						Repair.RepairFiles(outputFolder,config);
+						return Task.CompletedTask;
+					default:
+						Console.WriteLine($"{args[0]} is not a recognized argument. Exiting now.");
+						return Task.CompletedTask;
+				}
+				break;
+			case 2:
+				switch (args[0])
+				{
+					case "format" or "Format":
+						string newfmt = args[1];
+
+						if(Regex.Match("", @"^\.?(?:Wav|WAV|wav)$").Success)
+						{
+							config = Config.FromFile();
+							config.Format = AudioFormat.Wav;
+							Console.WriteLine("Changed format to .wav");
+							config.WriteToConfigFile();
+							return Task.CompletedTask;
+						}
+						else if(Regex.Match("", @"^\.?(?:Mp3|MP3|mp3)$").Success)
+						{
+							config = Config.FromFile();
+							config.Format = AudioFormat.Wav;
+							Console.WriteLine("Changed format to .wav");
+							config.WriteToConfigFile();
+							return Task.CompletedTask;
+						}
+						else
+						{
+							Console.WriteLine($"{newfmt} is not a recognized argument. Exiting now.");
+							return Task.CompletedTask;
+
+						}
+					default:
+						Console.WriteLine($"{args[0]} is not a recognized argument. Exiting now.");
+						return Task.CompletedTask;
+				}
+			default:
+				config = Config.FromFile();
+				break;
 		}
-		else
-		{
-			id_info = ConfigToID();
-		}
+
 
 		CancelKeyPress += (_, args) =>
 		{
@@ -84,7 +133,7 @@ class Program
 
 		waveIn = new NAudio.Wave.WaveInEvent
 		{
-			DeviceNumber = id_info.mic_id,
+			DeviceNumber = config.Mic_ID,
 			BufferMilliseconds = 100,
 			WaveFormat = loopback.WaveFormat
 		};
@@ -96,11 +145,10 @@ class Program
 
 		start = DateTime.Now;
 
-		string iso = $"{start:yyyy-MM-ddTHH-mm-ss}";
+		iso = $"{start:yyyy-MM-ddTHH-mm-ss}";
 
 		inpath =Path.Combine(outputFolder, $".{iso}_mic.wav");
 		stereopath = Path.Combine(outputFolder,$".{iso}_stereo.wav");
-		finalpath = Path.Combine(outputFolder,$"{iso}.wav");
 
 		waveFile = new WaveFileWriter(inpath,loopback.WaveFormat);
 		stereoFile = new WaveFileWriter(stereopath, loopback.WaveFormat);
@@ -114,11 +162,11 @@ class Program
 		waveIn.StartRecording();
 		loopback.StartRecording();
 
-		return RunLoop(cts.Token);
+		return RunLoop(cts.Token, config);
 
 	}
 
-	static async Task RunLoop(CancellationToken cancellation)
+	static async Task RunLoop(CancellationToken cancellation, Config config)
 	{
 		try
 		{
@@ -143,7 +191,7 @@ class Program
 		}
 		catch (TaskCanceledException)
 		{
-			CloseAndWrite();
+			CloseAndWrite(config);
 		}
 		catch (Exception exception)
 		{
@@ -154,7 +202,7 @@ class Program
 	static string deltaToString() => (end - start).ToString(@"hh\:mm\:ss");
 
 
-	static void CloseAndWrite()
+	static void CloseAndWrite(Config config)
 	{
 		if (!isRecording) return;
 		isRecording = false;
@@ -178,22 +226,20 @@ class Program
 		silenceOut.Dispose();
 		silenceOut = null;
 
-		using (var reader1 = new AudioFileReader(inpath))
+		var saveAction = config.FormatToSaveAction();
+
+		var result = saveAction(inpath, stereopath, iso, outputFolder);
+
+		if (result.Success)
 		{
-			using (var reader2 = new AudioFileReader(stereopath))
-			{
-				var mixer = new MixingSampleProvider([reader1, reader2]);
-				WaveFileWriter.CreateWaveFile16(finalpath, mixer);
-			}
+			File.Delete(inpath);
+			File.Delete(stereopath);
+			Console.WriteLine($"\n\nSaved {deltaToString()} to {result.FinalPath}");
 		}
-
-
-		File.Delete(inpath);
-		File.Delete(stereopath);
-
-		Console.WriteLine();
-		Console.WriteLine();
-		Console.WriteLine($"Saved {deltaToString()} to {finalpath}");
+		else
+		{
+			Console.WriteLine("\n\n Recording failed. The recorded channels can be fixed by running the app with the \"Fix\" argument.");
+		}
 	}
 
 	static void waveIn_DataAvailable(object sender, WaveInEventArgs e)
@@ -214,141 +260,6 @@ class Program
 		}
 	}
 
-	static (int id,string device_name) Setup()
-	{
-		Directory.CreateDirectory(configFolder);
-
-		var inCount = NAudio.Wave.WaveInEvent.DeviceCount;
-
-		for (int i = 0; i < inCount; i++)
-		{
-			var caps = WaveInEvent.GetCapabilities(i);
-			Console.WriteLine($"{i} , {caps.ProductName}");
-		}
-
-		var ok = InputToDevice(NAudio.Wave.WaveInEvent.DeviceCount, out var device, out int mic_id);
-
-		if (!ok) userExit("Cancelled setting an input.");
-
-		File.WriteAllText(config, $"{device.ProductName}");
-		File.SetAttributes(config, File.GetAttributes(config) | FileAttributes.Hidden);
-
-		return (mic_id,device.ProductName);
-	}
-
-	static (int id, string device_name) ConfigToID()
-	{
-		if (!File.Exists(config))
-		{
-			return Setup();
-		}
-
-		var lines = File.ReadAllLines(config);
-
-		if (lines.Length != 1) userExit("Error, expected one line in the config file. Run setup to fix the malformed config file.");
-
-		string maybe_name = lines[0];
-
-		int device_id = -1;
-
-		var inCount = NAudio.Wave.WaveInEvent.DeviceCount;
-
-		for (int i = 0; i < inCount; i++)
-		{
-			var caps = WaveInEvent.GetCapabilities(i);
-			if (string.Equals(caps.ProductName, maybe_name))
-			{
-				device_id = i;
-				break;
-			}
-		}
-
-		if (device_id == -1)
-		{
-			Console.WriteLine($"The device stored in the config file is no longer recognized. That device was {maybe_name} and is not currently an available option. Did you disable it? The app will now continue to setup to pick a new input device.\n\n");
-			return Setup();
-		}
-
-		return (device_id,maybe_name);
-	}
-
-	static bool InputToDevice(int devCount, out WaveInCapabilities devicecaps, out int dev_id)
-	{
-		var rec_index = GetRecDevice(devCount);
-		var device = WaveInEvent.GetCapabilities(rec_index);
-
-		var ok = validateRecDevice(rec_index, device);
-
-		if (!ok)
-		{
-			devicecaps = default;
-			dev_id = -1;
-			return InputToDevice(devCount, out devicecaps, out dev_id);
-		}
-		else
-		{
-			devicecaps = device;
-			dev_id = rec_index;
-			return true;
-		}
-	}
-
-	static void userExit(string msg)
-	{
-		Console.WriteLine($"UUser aborted, reason: {msg}. Press any key to quit.");
-		Console.ReadKey();
-		Environment.Exit(0);
-	}
-	static bool validateRecDevice(int i, WaveInCapabilities device)
-	{
-		Console.WriteLine();
-		Console.WriteLine($"Selected {i} , {device.ProductName}; Is this correct? Y/N");
-
-		var result = Console.ReadKey();
-
-		if (result.Key == ConsoleKey.Y) return true;
-		else if (result.Key == ConsoleKey.N) return false;
-		else if (result.Key == ConsoleKey.Escape) userExit("User quit the device selection");
-
-		return validateRecDevice(i, device);
-	}
-
-	static int GetRecDevice(int max)
-	{
-		Console.WriteLine();
-		Console.WriteLine("Enter the number of the recording device to listen to:");
-		Console.WriteLine();
-		var rawinput = Console.ReadLine();
-
-		if (rawinput == "quit" || rawinput == "q")
-		{
-			userExit("User quit device selection.");
-		}
-		if (!int.TryParse(rawinput, out var int_result))
-		{
-			Console.WriteLine("Error, input an integer corresponding with a device. Try again.");
-			return GetRecDevice(max);
-		}
-
-		if (int_result > max - 1)
-		{
-			Console.WriteLine();
-			Console.WriteLine($"Error, number is too large. Enter a number between 0 and {max - 1}. Try again.");
-
-			return GetRecDevice(max);
-		}
-
-		if (int_result < 0)
-		{
-			Console.WriteLine();
-			Console.WriteLine($"Error, number must be positive. Enter a number between 0 and {max - 1}. Try again.");
-
-			return GetRecDevice(max);
-		}
-
-
-		return int_result;
-	}
